Añadidos los warnings como errores y su solucion con static cast

+++ Flappy Cloak/src/obstacle/obstacle.cpp
@@ -3,15 +3,15 @@
 Obstacle::Obstacle()
 {
 	gap = 90; //Player size * 3
-	recTop.x = GetScreenWidth() + 100;
-	recTop.y = 0;
-	recTop.width = GetScreenWidth() / 10;
-	recTop.height = (GetScreenHeight() / 2) - gap;
+	recTop.x = static_cast<float>(GetScreenWidth() + 100);
+	recTop.y = 0.0f;
+	recTop.width = static_cast<float>(GetScreenWidth() / 10);
+	recTop.height = static_cast<float>((GetScreenHeight() / 2) - gap);
 
-	recBot.width = GetScreenWidth() / 10;
-	recBot.height = (GetScreenHeight() / 2) + gap;
-	recBot.x = GetScreenWidth() + 100;
-	recBot.y = (GetScreenHeight() / 2) + gap;
+	recBot.width = static_cast<float>(GetScreenWidth() / 10);
+	recBot.height = static_cast<float>((GetScreenHeight() / 2) + gap);
+	recBot.x = static_cast<float>(GetScreenWidth() + 100);
+	recBot.y = static_cast<float>((GetScreenHeight() / 2) + gap);
 	
 	color = RED;
 }
@@ -30,8 +30,8 @@ void Obstacle::respawn()
 {
 	if (recTop.x<-100 && recBot.x<-100)
 	{
-		recTop.x = GetScreenWidth() + 100;
-		recBot.x = GetScreenWidth() + 100;
+		recTop.x = static_cast<float>(GetScreenWidth() + 100);
+		recBot.x = static_cast<float>(GetScreenWidth() + 100);
 	}
 }
 
@@ -43,15 +43,15 @@ void Obstacle::draw()
 
 void Obstacle::reset()
 {
-	recTop.x = GetScreenWidth() + 100;
-	recTop.y = 0;
-	recTop.width = GetScreenWidth() / 10;
-	recTop.height = (GetScreenHeight() / 2) - gap;
+	recTop.x = static_cast<float>(GetScreenWidth() + 100);
+	recTop.y = 0.0f;
+	recTop.width = static_cast<float>(GetScreenWidth() / 10);
+	recTop.height = static_cast<float>((GetScreenHeight() / 2) - gap);
 
-	recBot.width = GetScreenWidth() / 10;
-	recBot.height = (GetScreenHeight() / 2) + gap;
-	recBot.x = GetScreenWidth() + 100;
-	recBot.y = (GetScreenHeight() / 2) + gap;
+	recBot.width = static_cast<float>(GetScreenWidth() / 10);
+	recBot.height = static_cast<float>((GetScreenHeight() / 2) + gap);
+	recBot.x = static_cast<float>(GetScreenWidth() + 100);
+	recBot.y = static_cast<float>((GetScreenHeight() / 2) + gap);
 }
 
 Rectangle Obstacle::getRecTop()
